solved the issue with the long

+++ src/main/java/nl/techIT/techITeasy/repository/TelevisionRepository.java
@@ -3,7 +3,7 @@ package nl.techIT.techITeasy.repository;
 import nl.techIT.techITeasy.model.Television;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TelevisionRepository extends JpaRepository<Television, Integer> {
+public interface TelevisionRepository extends JpaRepository<Television, Long> {
 
 
     Iterable<Television> findAllByNameContainingIgnoreCase(String name);
